extraindo em módulos pt3
