+++ space-invaders/src/game/mod.rs
@@ -16,8 +16,12 @@ use boss::BossPlugin;
 use bullet::BulletPlugin;
 use common::AppState;
 use laser::LaserPlugin;
+pub use lives::resources::Lives;
 use lives::LivesPlugin;
+pub use lives::LIVES;
+pub use player::player_sprite;
 use player::PlayerPlugin;
+pub use score::resources::Score;
 use score::ScorePlugin;
 use systems::*;
 
@@ -28,29 +32,25 @@ pub struct GamePlugin;
 
 impl Plugin for GamePlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(Startup, (spawn_score, spawn_lives_hud))
-            .add_systems(
-                Update,
-                (
-                    update_score,
-                    update_lives,
-                    collide_bullets_with_aliens,
-                    collide_lasers_with_player,
-                    collide_projectiles_with_barriers,
-                    collide_bullets_with_boss,
-                )
-                    .run_if(in_state(AppState::InGame)),
+        app.add_systems(
+            Update,
+            (
+                collide_bullets_with_aliens,
+                collide_lasers_with_player,
+                collide_projectiles_with_barriers,
+                collide_bullets_with_boss,
             )
-            .add_plugins((
-                AlienPlugin,
-                LaserPlugin,
-                BossPlugin,
-                PlayerPlugin,
-                BarrierPlugin,
-                BulletPlugin,
-                ScorePlugin,
-                LivesPlugin,
-            ))
-            .add_systems(OnExit(AppState::GameOver), respawn_live_hud);
+                .run_if(in_state(AppState::InGame)),
+        )
+        .add_plugins((
+            AlienPlugin,
+            LaserPlugin,
+            BossPlugin,
+            PlayerPlugin,
+            BarrierPlugin,
+            BulletPlugin,
+            ScorePlugin,
+            LivesPlugin,
+        ));
     }
 }
