+++ space-invaders/src/game/player/mod.rs
@@ -3,9 +3,18 @@ mod systems;
 
 use bevy::prelude::*;
 use common::AppState;
+use components::Player;
 use systems::*;
 
 pub const PLAYER_Y_OFFSET: f32 = 100.;
+pub fn player_sprite() -> Sprite {
+    let player = Player::default();
+    Sprite {
+        color: player.stats.color(),
+        custom_size: Some(player.stats.size()),
+        ..default()
+    }
+}
 
 pub struct PlayerPlugin;
