+++ space-invaders/src/game/systems.rs
@@ -4,124 +4,15 @@ use super::barrier::components::Barrier;
 use super::boss::components::Boss;
 use super::boss::BOSS_POINTS;
 use super::bullet::components::Bullet;
-use super::components::{HitPoints, LivesHUD, ScoreHUD, Stats};
+use super::components::{HitPoints, Stats};
 use super::laser::components::Laser;
 use super::lives::resources::Lives;
 use super::player::components::Player;
 use super::score::resources::Score;
 use bevy::prelude::*;
 use bevy::sprite::collide_aabb;
-use common::events::EndGame;
-use common::AppState;
-
-const LIVES_X_OFFSET: f32 = 30.;
-const LIVES_COL_OFFSET: f32 = 50.;
-const FONT_SIZE: f32 = 40.;
 
 type ProjectileQuery = Or<(With<Bullet>, With<Laser>)>;
-
-pub fn spawn_score(mut commands: Commands, asset_server: Res<AssetServer>, score: Res<Score>) {
-    const X_OFFSET: f32 = 500.;
-    commands.spawn((
-        Text2dBundle {
-            text: Text::from_section(
-                format!("Score {:05}", score.score()),
-                TextStyle {
-                    font: asset_server.load("fonts/FiraSans-Bold.ttf"),
-                    font_size: FONT_SIZE,
-                    color: Color::WHITE,
-                },
-            )
-            .with_alignment(TextAlignment::Center),
-            transform: Transform::from_xyz(X_OFFSET, FONT_SIZE / 2., 0.),
-            ..default()
-        },
-        ScoreHUD,
-    ));
-}
-
-pub fn update_score(score: Res<Score>, mut score_hud_query: Query<&mut Text, With<ScoreHUD>>) {
-    if let Ok(mut text) = score_hud_query.get_single_mut() {
-        for section in &mut text.sections {
-            section.value = format!("Score {:05}", score.score());
-        }
-    }
-}
-
-pub fn spawn_lives_hud(mut commands: Commands, lives: Option<Res<Lives>>) {
-    let player = Player::default();
-    let sprite = Sprite {
-        color: player.stats.color(),
-        custom_size: Some(player.stats.size()),
-        ..default()
-    };
-    if let Some(lives) = lives {
-        for i in 0..lives.get() {
-            commands.spawn((
-                SpriteBundle {
-                    sprite: sprite.clone(),
-                    transform: Transform::from_xyz(
-                        LIVES_X_OFFSET + i as f32 * LIVES_COL_OFFSET,
-                        FONT_SIZE / 2.,
-                        0.0,
-                    ),
-                    ..default()
-                },
-                LivesHUD,
-            ));
-        }
-    }
-}
-
-pub fn update_lives(
-    lives: Option<Res<Lives>>,
-    score_hud_query: Query<Entity, With<LivesHUD>>,
-    mut commands: Commands,
-    mut game_over_event_writer: EventWriter<EndGame>,
-    score: Res<Score>,
-) {
-    if let Some(lives) = lives {
-        if lives.is_changed() && !lives.is_added() {
-            let len = score_hud_query.iter().len();
-            let lives = lives.get();
-            if lives == 0 {
-                for entity in score_hud_query.iter() {
-                    commands.entity(entity).despawn();
-                }
-                commands.insert_resource(NextState(Some(AppState::GameOver)));
-                game_over_event_writer.send(EndGame {
-                    score: score.score(),
-                });
-            } else if len - 1 == lives {
-                if let Some(entity) = score_hud_query.iter().last() {
-                    commands.entity(entity).despawn();
-                }
-            // É importante checar se a variação é realmente 1 pois ao resetar o contador de vidas
-            // (isto é, mudar de 0 para 3), a variação é maior que 1 e não queremos spawnar um
-            // novo sprite porque o reset toma conta disso
-            } else if len + 1 == lives {
-                let player = Player::default();
-                let sprite = Sprite {
-                    color: player.stats.color(),
-                    custom_size: Some(player.stats.size()),
-                    ..default()
-                };
-                commands.spawn((
-                    SpriteBundle {
-                        sprite: sprite.clone(),
-                        transform: Transform::from_xyz(
-                            LIVES_X_OFFSET + len as f32 * LIVES_COL_OFFSET,
-                            FONT_SIZE / 2.,
-                            0.0,
-                        ),
-                        ..default()
-                    },
-                    LivesHUD,
-                ));
-            }
-        }
-    }
-}
 
 pub fn collide_bullets_with_aliens(
     mut commands: Commands,
@@ -221,15 +112,4 @@ pub fn collide_bullets_with_boss(
             }
         }
     }
-}
-
-pub fn respawn_live_hud(
-    mut commands: Commands,
-    hud_query: Query<Entity, With<LivesHUD>>,
-    lives: Option<Res<Lives>>,
-) {
-    for entity in hud_query.iter() {
-        commands.entity(entity).despawn();
-    }
-    spawn_lives_hud(commands, lives);
 }
