+++ space-invaders/src/hud/lives/components.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component)]
+pub struct LivesHUD;
