+++ space-invaders/src/hud/score/components.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component)]
+pub struct ScoreHUD;
